+++ src/components/SchematicDwg.js
@@ -1,0 +1,9 @@
+/*jshint esversion: 6 */
+import React from "react";
+
+const SchematicDwg = ({ schematic }) => {
+  console.log(this.state.position);
+  return <img src={schematic} />;
+};
+
+export default SchematicDwg;
