+++ src/containers/Schematic/Schematic.js
@@ -1,10 +1,25 @@
 /*jshint esversion: 6 */
 import React, { Component } from "react";
+import axios from "axios";
 import ReactCursorPosition from "react-cursor-position";
+import Modal from "react-modal";
+import ModalContent from "../../containers/ModalContent/ModalContent.js";
 import Header from "../Header/Header.js";
 import Footer from "../../components/Footer.js";
-import SchematicDwg from "./SchematicDwg";
+import SchematicDwg from "../../components/SchematicDwg.js";
 import { getProjectById } from "../../lib/projects-api";
+import Pin from "../../containers/Pin/Pin.js";
+
+const customStyles = {
+  content: {
+    top: "50%",
+    left: "50%",
+    right: "auto",
+    bottom: "auto",
+    marginRight: "-50%",
+    transform: "translate(-50%, -50%)"
+  }
+};
 
 class Schematic extends Component {
   constructor(props) {
@@ -13,14 +28,14 @@ class Schematic extends Component {
     this.state = {
       project: null,
       schematic: null,
-      pins: []
+      pins: [],
+      activate: false,
+      modalIsOpen: false
     };
   }
 
   componentDidMount() {
     getProjectById(this.projectId).then(projectData => {
-      console.log("project data", projectData[0]);
-      console.log("project data", projectData[0].pin);
       localStorage.setItem("projectId", projectData[0].project.id);
       localStorage.setItem("schematicId", projectData[0].schematic.id);
       let schematicId = localStorage.getItem("schematicId");
@@ -32,32 +47,94 @@ class Schematic extends Component {
     });
   }
 
+  activatePins() {
+    this.setState({
+      activate: true
+    });
+  }
+
+  logPostition() {
+    // if (this.state.activate === true) {
+    //   let username = localStorage.getItem("loggedInUserName");
+    //   let userId = localStorage.getItem("loggedInUserId");
+    //   let schematicId = localStorage.getItem("schematicId");
+    //   let projectId = localStorage.getItem("projectId");
+    //   // console.log("projectId", projectId);
+    //   // console.log("schematicId", schematicId);
+    //   // console.log("username =", username);
+    //   // console.log("userId", userId);
+    //   console.log(this.props.position);
+    //   // console.log("x", this.state.position.x);
+    //   // console.log("y", this.state.position.y);
+    //   let newPin = {
+    //     x: this.state.position.x,
+    //     y: this.state.position.y,
+    //     user_id: userId,
+    //     project_id: projectId,
+    //     schematic_id: schematicId
+    //   };
+    //   axios
+    //     .post("/api/pins", newPin)
+    //     .then(function(response) {
+    //       console.log("new pin dropped", response);
+    //     })
+    //     .catch(function(error) {
+    //       console.log(error);
+    //     });
+    //   this.setState({
+    //     pinList: [...this.state.pinList, this.state.position],
+    //     activate: false
+    //   });
+    // } else {
+    //   console.log("hit the button");
+    // }
+  }
+
+  openModal() {
+    this.setState({ modalIsOpen: true });
+  }
+
+  closeModal() {
+    this.setState({ modalIsOpen: false });
+  }
+
   render() {
+    console.log(this.state.project);
     console.log(this.state.pins);
-    if (this.state.pins.length > 0) {
-      return (
-        <div>
-          <Header />
-          <ReactCursorPosition>
-            <SchematicDwg
-              image={
-                this.state.schematic
-                  ? this.state.schematic.image_url
-                  : "Image did not load"
-              }
-              pinList={this.state.pins}
-            />
-          </ReactCursorPosition>
-          <Footer
-            project={
-              this.state.project ? this.state.project : "Props did not load"
+    console.log("schematic rendering");
+
+    const pinLayer = {
+      position: "relative",
+      width: "100%",
+      height: "100%"
+    };
+
+    return (
+      <div>
+        <Header />
+        <div className="pinLayer" style={pinLayer}>
+          {this.state.pins.map(pin => {
+            return <Pin x={pin.x} y={pin.y} />;
+          })}
+        </div>
+        <ReactCursorPosition>
+          <SchematicDwg
+            schematic={
+              this.state.schematic
+                ? this.state.schematic.image_url
+                : "Props did not load"
             }
           />
-        </div>
-      );
-    } else {
-      return <div>stupid pins</div>;
-    }
+          <button onClick={this.activatePins.bind(this)}>Drop Pins</button>
+        </ReactCursorPosition>
+
+        <Footer
+          project={
+            this.state.project ? this.state.project : "Props did not load"
+          }
+        />
+      </div>
+    );
   }
 }
